requirements.txt
Flask
python-pptx
google-generativeai
